# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Authway login UI API
  version: 0.5.0
  summary: The contract between Authway's login screens and a provider.
  description: |
    Every Authway implementation answers these endpoints the same way, so the
    bundled login screens — or a custom one — work with any of them: sign-in,
    consent and logout flows, email verification, password reset and
    invitations. Flow ids
    are opaque: a screen only passes them back, URL-encoded in a path segment.
    See docs/api/login-flows.md for the walk-through.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
security: []
paths:
...
  /api/v1/webhooks:
    get:
      tags: [admin]
      operationId: listWebhooks
      security: [{ adminKey: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }]
      responses:
        '200':
          description: The tenant's webhooks, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [webhooks]
                properties:
                  webhooks:
                    type: array
                    items: { $ref: '#/components/schemas/Webhook' }
        '400':
          description: No tenant named (code tenant_required)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [admin]
      operationId: createWebhook
      description: |
        Registers an endpoint for the tenant's events. A value outside its
        range is refused (code invalid_request), never replaced; an omitted
        optional value takes its default.
      security: [{ adminKey: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url, events]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                url:
                  type: string
                  description: |
                    An absolute http or https URL. A local or private address is
                    refused unless the deployment allows private targets.
                events:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/WebhookEvent' }
                enabled: { type: boolean, default: true }
                retry_count: { type: integer, minimum: 0, maximum: 10, default: 3 }
                timeout_secs: { type: integer, minimum: 1, maximum: 60, default: 30 }
      responses:
        '201':
          description: Created. The only answer that carries the signing secret.
          content:
            application/json:
              schema:
                type: object
                required: [webhook, secret]
                properties:
                  webhook: { $ref: '#/components/schemas/Webhook' }
                  secret:
                    type: string
                    description: Signs every delivery; see docs/api/webhooks.md. Not shown again — rotate to replace it.
                  message: { type: string }
        '400':
          description: Invalid value (code invalid_request), or no tenant named (code tenant_required)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/webhooks/events:
    get:
      tags: [admin]
      operationId: listWebhookEvents
      security: [{ adminKey: [] }]
      responses:
        '200':
          description: What a webhook can subscribe to
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      type: object
                      required: [type, description]
                      properties:
                        type: { $ref: '#/components/schemas/WebhookEvent' }
                        description: { type: string }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/webhooks/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [admin]
      operationId: getWebhook
      security: [{ adminKey: [] }]
      responses:
        '200':
          description: The webhook
          content:
            application/json:
              schema:
                type: object
                required: [webhook]
                properties:
                  webhook: { $ref: '#/components/schemas/Webhook' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [admin]
      operationId: updateWebhook
      description: Changes the fields the body names; the same ranges as creation apply.
      security: [{ adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                url: { type: string }
                events:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/WebhookEvent' }
                enabled: { type: boolean }
                retry_count: { type: integer, minimum: 0, maximum: 10 }
                timeout_secs: { type: integer, minimum: 1, maximum: 60 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [webhook]
                properties:
                  webhook: { $ref: '#/components/schemas/Webhook' }
                  message: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [admin]
      operationId: deleteWebhook
      security: [{ adminKey: [] }]
      responses:
        '200': { $ref: '#/components/responses/Message' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/webhooks/{id}/test:
    post:
      tags: [admin]
      operationId: testWebhook
      description: |
        Sends one test event to this webhook — enabled or not, whatever it
        subscribes to — and answers with the delivery. A receiver that refuses
        or cannot be reached still answers 200; the delivery says how it went.
      security: [{ adminKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: The delivery
          content:
            application/json:
              schema:
                type: object
                required: [delivery]
                properties:
                  delivery: { $ref: '#/components/schemas/WebhookDelivery' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/webhooks/{id}/rotate-secret:
    post:
      tags: [admin]
      operationId: rotateWebhookSecret
      description: Replaces the signing secret; the previous one stops signing at once.
      security: [{ adminKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: The new secret, shown this once
          content:
            application/json:
              schema:
                type: object
                required: [secret]
                properties:
                  secret: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    adminKey:
      type: http
      scheme: bearer
      description: The deployment's admin API key, or an admin console session token.
  parameters:
...
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: The tenant the call acts on. A tenant_id query parameter works too.
      schema: { type: string }
...
  responses:
    Message:
      description: Done
      content:
        application/json:
          schema:
            type: object
            required: [message]
            properties:
              message: { type: string }
...
    Error:
      description: Refused
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Webhook:
      type: object
      required: [id, tenant_id, name, url, events, enabled, retry_count, timeout_secs, created_at, updated_at]
      properties:
        id: { type: string }
        tenant_id: { type: string }
        name: { type: string }
        url: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/WebhookEvent' }
        enabled: { type: boolean }
        retry_count: { type: integer, minimum: 0, maximum: 10, description: Attempts after the first. }
        timeout_secs: { type: integer, minimum: 1, maximum: 60 }
        created_at: { type: string }
        updated_at: { type: string }
    WebhookEvent:
      type: string
      enum: [user.created, user.updated, user.deleted, user.login, user.logout, user.password_changed,
             user.mfa_enabled, user.mfa_disabled, session.created, session.revoked,
             client.created, client.updated, client.deleted, test, '*']
    WebhookDelivery:
      type: object
      required: [id, webhook_id, event_type, payload, status_code, attempt, delivered_at, success]
      properties:
        id: { type: string }
        webhook_id: { type: string }
        event_type: { type: string }
        payload: { type: string, description: The JSON body that was sent. }
        status_code: { type: integer, description: 0 when no answer arrived. }
        response_body: { type: string }
        attempt: { type: integer, minimum: 1 }
        delivered_at: { type: string }
        success: { type: boolean }
        error_message: { type: string }
...
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: A message to show the user. }
        code:
          type: string
          description: |
            Stable identifier, for example invalid_flow, flow_expired,
            invalid_credentials, sign_in_method_not_allowed,
            invalid_mfa_challenge, invalid_code, mfa_locked, invalid_link,
            invalid_request, client_not_registered, invalid_token,
            token_expired, invitation_not_acceptable, invitation_not_accepted,
            too_many_requests, not_found, method_not_allowed, unauthorized,
            forbidden, insufficient_scope, admin_api_not_configured, conflict,
            public_client_has_no_secret, upstream_sync_failed,
            tenant_slug_taken, default_tenant_protected, tenant_has_users,
            tenant_has_clients, tenant_has_service_clients, tenant_required,
            user_already_member, invitation_already_pending,
            invitation_not_pending,
            internal_server_error. A refusal that carries only an HTTP status
            uses the code named after it.